+++ test/expression/literal_operators.js
@@ -1,5 +1,5 @@
 var test = require('tape')
-var mingo = require('../dist/mingo')
+var mingo = require('../../dist/mingo')
 
 test('Literal Operators', function (t) {
   t.plan(1)